Refactoring and updating methods. Ready for a 1.1-BETA
