+++ README.md
@@ -7,7 +7,7 @@ This toolkit will be used throughout Londogard libraries/products such as our Su
 
 First of, `LanguageSupport.<ISO_COUNTRY_CODE>`helps figuring out what is and isn't supported for certain languages. Utilities that makes sense to provide your own data, like Word Embeddings & Tokenizers, are available for any language if you provide your own models.
 
-| Feature                     | Implementations (if multiple)                                             | Languages                                                                                                                                 | Based On                                                                                                                                                | Notes                                                                                                                                                                                                                                                                                      |
+| Feature                     | Variants (if multiple)                                             | Languages                                                                                                                                 | Based On                                                                                                                                                | Notes                                                                                                                                                                                                                                                                                      |
 |-----------------------------|---------------------------------------------------------------------------|-------------------------------------------------------------------------------------------------------------------------------------------|---------------------------------------------------------------------------------------------------------------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | **Word Embeddings**         | - `WordEmbeddings`<br>- `LightWordEmbeddings`<br>- `BPEEmbeddings`        | - ∞ languages if self-provided<br>- 275 through `BPEEmb`-embeddings<br>- 157 through `fastText`-embeddings                                | - `fastText` embeddings downloaded from [fastText.cc](https://fasttext.cc)<br>- `bpe-emb` embeddings downloaded from [bpeemb](https://bpemb.h-its.org/) | - `LightWordEmbeddings`is great for memory (e.g. running on Raspberry Pi) as it only loads the most common embeddings and cache ad-hoc<br>- `BPEEmbeddings` are also great for memory but requires subword tokenization (built-in so you don't have to think about it unless you optimize) |
 | **Sentence Embeddings**     | - `AverageSentenceEmbeddings`<br>- `USifSentenceEmbeddings`               |                                                                                                                                           | USif is based on [Unsupervised Random Walk Sentence Embeddings: A Strong but Simple Baseline](https://aclanthology.org/W18-3012/)                       |                                                                                                                                                                                                                                                                                            |
