+++ src/main/kotlin/com/londogard/nlp/meachinelearning/datatypes/Coordinate.kt
@@ -1,3 +1,3 @@
-package com.londogard.nlp.meachinelearning.inputs
+package com.londogard.nlp.meachinelearning.datatypes
 
 data class Coordinate<T: Number>(val row: Int, val col: Int, val count: T)
