--- src/main/kotlin/com/londogard/nlp/simple/NLP.kt
@@ -1,8 +1,0 @@
-package com.londogard.nlp.simple
-
-object NLP {
-    var pipeline: Pipeline = TODO("")
-
-
-    fun consume(text: String): Document = TODO("")
-}
