--- src/main/kotlin/com/londogard/nlp/simple/Pipeline.kt
@@ -1,8 +1,0 @@
-package com.londogard.nlp.simple
-
-interface Pipeline {
-    val preprocessing: List<Stage>
-    val tokenizing: Stage
-    val preprocessTokens: List<Stage>
-    val classifiers: List<Stage>
-}
