--- src/main/kotlin/com/londogard/nlp/simple/Stage.kt
@@ -1,3 +1,0 @@
-package com.londogard.nlp.simple
-
-interface Stage { }
